feat: 여행지 목록, 상세조회에 좋아요 여부, 좋아요 개수 추가

+++ src/main/java/com/ssafy/trip/domain/like/controller/LikeController.java
@@ -23,7 +23,7 @@ public class LikeController {
     }
 
     @GetMapping
-    public ResponseEntity count(@RequestParam Long spotId) throws Exception {
+    public ResponseEntity count(@RequestParam int spotId) throws Exception {
         int cnt = likeService.count(spotId);
         return ResponseEntity.ok(Response.success(cnt));
     }
